+++ Automate/release-notes.md
@@ -1,6 +1,9 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.10.6
+* Migrated verbose logs to SMAPI's verbose logging option.
+
 ## 1.10.5
 * Updated for SMAPI 2.8-beta.5.
 * Fixed error if a mill contains an unrecognised item.